+++ packages-serverless/runtime-engine/test/lib/logger.test.ts
@@ -1,9 +1,7 @@
 import * as assert from 'assert';
 import * as os from 'os';
 import * as path from 'path';
-
-import { ServerlessLogger } from '../../src/lib/logger';
-import { BaseLoggerFactory } from '../../src/lib/loggerFactory';
+import { ServerlessLogger, BaseLoggerFactory } from '../../src';
 
 describe('logger.test.ts', () => {
   describe('logger', () => {
@@ -18,6 +16,7 @@ describe('logger.test.ts', () => {
         consoleLevel: 'NONE',
         buffer: true,
       });
+      logger.close();
     });
 
     it('should logger with file', () => {
@@ -25,13 +24,14 @@ describe('logger.test.ts', () => {
         file: path.join(os.tmpdir(), 'test.log'),
       });
       logger.write('hello');
+      logger.close();
     });
 
-    it('should do logger file clear without no op', () => {
+    it('should do logger file clear without no op', async () => {
       class MyLogger extends ServerlessLogger {
-        test() {
+        async test() {
           this.options.fileClearInterval = 10;
-          this.rotateLogBySize();
+          await this.rotateLogBySize();
         }
       }
       const logger = new MyLogger({
@@ -39,15 +39,16 @@ describe('logger.test.ts', () => {
         level: 'ALL',
       });
       logger.write('hello, world!');
-      logger.test();
+      await logger.test();
+      logger.close();
     });
 
-    it('should do logger file clear', () => {
+    it('should do logger file clear', async () => {
       class MyLogger extends ServerlessLogger {
-        test() {
+        async test() {
           this.options.fileClearInterval = 10;
           this.options.maxFileSize = 5;
-          this.rotateLogBySize();
+          await this.rotateLogBySize();
         }
       }
       const logger = new MyLogger({
@@ -55,63 +56,68 @@ describe('logger.test.ts', () => {
         level: 'INFO',
       });
       logger.write('hello, world!');
-      logger.test();
+      await logger.test();
+      logger.close();
     });
 
-    it('should do logger file clear without log file', () => {
+    it('should do logger file clear without log file', async () => {
       class MyLogger extends ServerlessLogger {
-        test() {
+        async test() {
           this.options.fileClearInterval = 10;
           this.options.maxFileSize = 5;
-          this.rotateLogBySize();
+          await this.rotateLogBySize();
         }
       }
       const logger = new MyLogger({
         level: 'INFO',
       });
       logger.write('hello, world!');
-      logger.test();
+      await logger.test();
+      logger.close();
     });
 
-    it('should do logger rotateBySize without log file', () => {
+    it('should do logger rotateBySize without log file', async () => {
       class MyLogger extends ServerlessLogger {
-        test() {
+        async  test() {
           this.options.file = '';
           this.options.fileClearInterval = 10;
           this.options.maxFileSize = 5;
-          this.rotateBySize();
+          await this.rotateBySize();
         }
       }
       const logger = new MyLogger({
         level: 'INFO',
       });
       logger.write('hello, world!');
-      logger.test();
+      await logger.test();
+      logger.close();
     });
 
-    it('should do logger renameOrDelete without log file', () => {
+    it('should do logger renameOrDelete without log file', async () => {
       class MyLogger extends ServerlessLogger {
-        test() {
+        async test() {
           this.options.file = '';
           this.options.fileClearInterval = 10;
           this.options.maxFileSize = 5;
-          this.renameOrDelete('', '');
+          await this.renameOrDelete('', '');
         }
       }
       const logger = new MyLogger({
         level: 'INFO',
       });
       logger.write('hello, world!');
-      logger.test();
+      await logger.test();
+      logger.close();
     });
   });
 
   describe('loggerFactory', () => {
     it('should new loggerFactory', () => {
       const loggerFactory = new BaseLoggerFactory(__dirname);
-      loggerFactory.createLogger({
+      const logger = loggerFactory.createLogger({
         file: path.join(__dirname, 'fixtures'),
       });
+      logger.close();
     });
   });
 
@@ -132,6 +138,7 @@ describe('logger.test.ts', () => {
         //   setTimeout(() => resolve(), 50);
         // });
       }
+      log.close();
     });
   });
 });