+++ packages-serverless/serverless-fc-starter/src/runtime.ts
@@ -10,10 +10,13 @@ import {
 } from '@midwayjs/serverless-http-parser';
 import * as util from 'util';
 
-const isLocalEnv = () => {
+const isOutputError = () => {
   return (
-    process.env.MIDWAY_SERVER_ENV === 'local' ||
-    process.env.NODE_ENV === 'local'
+    process.env.SERVERLESS_OUTPUT_ERROR_STACK === 'true' ||
+    ['local', 'development'].includes(
+      process.env.MIDWAY_SERVER_ENV
+    ) ||
+    ['local', 'development'].includes(process.env.NODE_ENV)
   );
 };
 
@@ -201,13 +204,13 @@ export class FCRuntime extends ServerlessLightRuntime {
             ctx.logger.error(err);
             if (res.send) {
               res.setStatusCode(500);
-              res.send(isLocalEnv() ? err.stack : 'Internal Server Error');
+              res.send(isOutputError() ? err.stack : 'Internal Server Error');
             }
             return {
               isBase64Encoded: false,
               statusCode: 500,
               headers: {},
-              body: isLocalEnv() ? err.stack : 'Internal Server Error',
+              body: isOutputError() ? err.stack : 'Internal Server Error',
             };
           });
       },