+++ packages/core/src/util/webRouterCollector.ts
@@ -8,6 +8,7 @@ import {
   getPropertyDataFromClass,
   getPropertyMetadata,
   getProviderId,
+  isRegExp,
   listModule,
   PRIORITY_KEY,
   RouterOption,
@@ -99,7 +100,7 @@ export interface RouterCollectorOptions {
 export class WebRouterCollector {
   protected readonly baseDir: string;
   private isReady = false;
-  private routes = new Map<string, RouterInfo[]>();
+  protected routes = new Map<string, RouterInfo[]>();
   private routesPriority: RouterPriority[] = [];
   protected options: RouterCollectorOptions;
 
@@ -213,7 +214,7 @@ export class WebRouterCollector {
           } as FaaSMetadata.HTTPTriggerOptions;
         }
 
-        this.routes.get(prefix).push(data);
+        this.checkDuplicateAndPush(prefix, data);
       }
     }
   }
@@ -264,7 +265,7 @@ export class WebRouterCollector {
               module,
               webRouter['methodName']
             ) || [];
-          // 新 http/apigateway 函数
+          // 新 http/api gateway 函数
           const data: RouterInfo = {
             prefix,
             routerName: '',
@@ -286,11 +287,11 @@ export class WebRouterCollector {
             data.functionTriggerName = webRouter['type'];
             data.functionTriggerMetadata = webRouter['metadata'];
           }
-          this.routes.get(prefix).push(data);
+          this.checkDuplicateAndPush(prefix, data);
         } else {
           if (functionMeta) {
             // 其他类型的函数
-            this.routes.get(prefix).push({
+            this.checkDuplicateAndPush(prefix, {
               prefix,
               routerName: '',
               url: '',
@@ -341,11 +342,11 @@ export class WebRouterCollector {
             };
           }
           // 老函数的 http
-          this.routes.get(prefix).push(data);
+          this.checkDuplicateAndPush(prefix, data);
         } else {
           if (functionMeta) {
             // 非 http
-            this.routes.get(prefix).push({
+            this.checkDuplicateAndPush(prefix, {
               prefix,
               routerName: '',
               url: '',
@@ -381,6 +382,19 @@ export class WebRouterCollector {
     return urlMatchList
       .map(item => {
         const urlString = item.url.toString();
+        const weightArr = isRegExp(item.url)
+          ? urlString.split('/')
+          : urlString.split('/');
+        let weight = 0;
+        // 权重，比如通配的不加权，非通配加权，防止通配出现在最前面
+        for (const fragment of weightArr) {
+          if (fragment.includes(':') || fragment.includes('*')) {
+            weight += 0;
+          } else {
+            weight += 1;
+          }
+        }
+
         let category = 2;
         const paramString = urlString.includes(':')
           ? urlString.replace(/:.+$/, '')
@@ -397,6 +411,7 @@ export class WebRouterCollector {
           _level: urlString.split('/').length - 1,
           _paramString: paramString,
           _category: category,
+          _weight: weight,
         };
       })
       .sort((handlerA, handlerB) => {
@@ -404,7 +419,13 @@ export class WebRouterCollector {
         if (handlerA._category !== handlerB._category) {
           return handlerB._category - handlerA._category;
         }
+        // 不同长度
         if (handlerA._level === handlerB._level) {
+          // 不同权重
+          if (handlerA._weight !== handlerB._weight) {
+            return handlerB._weight - handlerA._weight;
+          }
+
           if (handlerB._pureRouter === handlerA._pureRouter) {
             return (
               handlerA.url.toString().length - handlerB.url.toString().length
@@ -442,5 +463,23 @@ export class WebRouterCollector {
       routeArr = routeArr.concat(routerInfo);
     }
     return routeArr;
+  }
+
+  private checkDuplicateAndPush(prefix, routerInfo: RouterInfo) {
+    const prefixList = this.routes.get(prefix);
+    const matched = prefixList.filter(item => {
+      return (
+        routerInfo.url &&
+        routerInfo.requestMethod &&
+        item.url === routerInfo.url &&
+        item.requestMethod === routerInfo.requestMethod
+      );
+    });
+    if (matched && matched.length) {
+      throw new Error(
+        `Duplicate router "${routerInfo.requestMethod} ${routerInfo.url}" at "${matched[0].handlerName}" and "${routerInfo.handlerName}"`
+      );
+    }
+    prefixList.push(routerInfo);
   }
 }