+++ packages/core/test/fixtures/app-with-duplicate-router/src/config/config.prod.ts
@@ -1,0 +1,1 @@
+// load when NODE_ENV=production