+++ packages/core/test/util/webRouterCollector.test.ts
@@ -1,13 +1,13 @@
-import { WebRouterCollector } from '../src';
+import { WebRouterCollector } from '../../src';
 import { join } from 'path';
 import { clearAllModule } from '@midwayjs/decorator';
-import { clearContainerCache } from '../src';
-import { matchObjectPropertyInArray } from './util';
+import { clearContainerCache } from '../../src';
+import { matchObjectPropertyInArray } from '../util';
 
-describe('/test/webRouterCollector.test.ts', function () {
+describe('/test/util/webRouterCollector.test.ts', function () {
 
   it('should test generate router', async () => {
-    const collector = new WebRouterCollector(join(__dirname, './fixtures/base-app-controller'));
+    const collector = new WebRouterCollector(join(__dirname, '../fixtures/base-app-controller'));
     const result = await collector.getRouterTable();
     expect(result.size).toEqual(3);
 
@@ -16,7 +16,7 @@ describe('/test/webRouterCollector.test.ts', function () {
   });
 
   it('should test generate flatten router', async () => {
-    const collector = new WebRouterCollector(join(__dirname, './fixtures/base-app-controller'));
+    const collector = new WebRouterCollector(join(__dirname, '../fixtures/base-app-controller'));
     const result = await collector.getFlattenRouterTable();
     expect(result.length > 0).toBeTruthy();
   });
@@ -24,7 +24,7 @@ describe('/test/webRouterCollector.test.ts', function () {
   it('should test with function router', async () => {
     clearAllModule();
     clearContainerCache();
-    const collector = new WebRouterCollector(join(__dirname, './fixtures/base-app-func-router'), { includeFunctionRouter: true});
+    const collector = new WebRouterCollector(join(__dirname, '../fixtures/base-app-func-router'), { includeFunctionRouter: true});
     const result = await collector.getFlattenRouterTable();
     expect(result.length).toEqual(6);
     expect(matchObjectPropertyInArray(result, {
@@ -98,4 +98,18 @@ describe('/test/webRouterCollector.test.ts', function () {
     expect(result[1].url).toEqual('/');
     expect(result[2].url).toEqual('/*');
   });
+
+  it('fix issue 1008', function () {
+    const collector = new WebRouterCollector();
+    const result1 = collector.sortRouter(require('./router').routerList3);
+    expect(result1[0].url).toEqual('/:page/page');
+    expect(result1[1].url).toEqual('/page/:page');
+    expect(result1[2].url).toEqual('/:category/:slug');
+
+    const result2 = collector.sortRouter(require('./router').routerList4);
+    expect(result2[0].url).toEqual('/page/:page');
+    expect(result2[1].url).toEqual('/:page/page');
+    expect(result2[2].url).toEqual('/:category/:slug');
+  });
+
 });