+++ packages/web/test/fixtures/feature/base-app-egg-logger/src/configuration.ts
@@ -1,0 +1,22 @@
+import { Configuration, Logger, App } from '@midwayjs/decorator';
+import * as assert from 'assert';
+import { Application } from 'egg';
+
+@Configuration({
+  importConfigs: [
+    './config'
+  ]
+})
+export class ContainerConfiguration {
+
+  @Logger()
+  logger;
+
+  @App()
+  app: Application;
+
+  async onReady() {
+    assert(this.logger.get('file'));
+    assert(this.logger === this.app.logger);
+  }
+}