+++ packages-serverless/runtime-mock/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.10.12](https://github.com/midwayjs/midway-faas/compare/v2.10.11...v2.10.12) (2021-05-07)
+
+**Note:** Version bump only for package @midwayjs/runtime-mock
+
+
+
+
+
+## [2.10.10](https://github.com/midwayjs/midway-faas/compare/v2.10.9...v2.10.10) (2021-04-24)
+
+**Note:** Version bump only for package @midwayjs/runtime-mock
+
+
+
+
+
 ## [2.10.7](https://github.com/midwayjs/midway-faas/compare/v2.10.6...v2.10.7) (2021-04-17)